requirements.txt
requests
python-dotenv
datasets
transformers
flask
flask-cors
